requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
httpx==0.28.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pytest==8.3.2
